Continued Header functions
Started off by looping throughs chars with forEach passing in char and index to use.  Inside that I run handleMouseEnter and handleMouseEnter  which changes events based on mouse events. When thew mouse enters the letter it runs scaleChange and when it leaves it runs resetTransformation. ScaleChange checks where the index is and transforms the matching letter to a larger size. ResetTransform simply reverts them when the mouse leaves the letter. This gives the whole element a cool wavey scale up and down affect.