Created Projects section
This section of the Home page is structured in a grid format, showcasing six of my projects. Each project is represented by a ProjectTab component that includes all the relevant information. I’m dynamically updating the details for each project by plugging in the specific content. Additionally, I’ve added a subtle GSAP animation for the text, making it slide in smoothly. Since this section is positioned relative and the header is sticky, it creates a cool sliding effect when scrolling down to this part of the page.