Created DisplayProject
This Page includes a more in depth look at each project. It showcases information like the name, languages used, and a description. Also displaying photos of the website. The info for which project to display is passed through the Link in ProjectTabs with each state.(Still need to add images. Most likely will use a database)