Revised ProjectTabs and created ProjectData
Decided it would be easier to use a json file for content and map through it so created ProjectData which holds all six projects. ProjectTabs works pretty similar except for mapping through the data and making it so I don't have to set the states I just pass them through to DisplayProject.