Created and added content to About section
This section is the last of the Home page and includes content about myself: the first part is my skills including the languages and techs I use, the second being and description about myself, and the final being social Links. Used two arrays for the skill boxes includes the icon and the name. It simply changes between them upon hover over and off using a ternary operator. Added About section to Home page.